move image generation to badge_utils
