+++ commands/badge_sets.py
@@ -4,6 +4,7 @@ import textwrap
 import time
 
 from common import *
+from utils.badge_utils import generate_paginated_badge_images
 
 from utils.check_channel_access import access_check
 
@@ -156,6 +157,14 @@ async def badge_sets(ctx:discord.ApplicationContext, public:str, category:str, s
   # If we're doing a public display, use the images directly
   # Otherwise private displays can use the paginator
   if not public:
+    buttons = [
+      pages.PaginatorButton("prev", label="   ⬅   ", style=discord.ButtonStyle.primary, row=1),
+      pages.PaginatorButton(
+        "page_indicator", style=discord.ButtonStyle.gray, disabled=True, row=1
+      ),
+      pages.PaginatorButton("next", label="   ➡   ", style=discord.ButtonStyle.primary, row=1),
+    ]
+
     set_pages = [
       pages.Page(files=[image], embeds=[embed])
       for image in set_images
@@ -164,6 +173,8 @@ async def badge_sets(ctx:discord.ApplicationContext, public:str, category:str, s
         pages=set_pages,
         show_disabled=True,
         show_indicator=True,
+        use_default_buttons=False,
+        custom_buttons=buttons,
         loop_pages=True
     )
     await paginator.respond(ctx.interaction, ephemeral=True)
@@ -185,7 +196,8 @@ async def generate_badge_set_images(user:discord.User, all_badges, title, collec
   all_pages = [all_badges[i:i + max_per_image] for i in range(0, len(all_badges), max_per_image)]
   total_pages = len(all_pages)
   badge_set_images = [
-    await generate_badge_set_showcase_for_user(
+    await generate_paginated_badge_images(
+      'sets', # type
       user_display_name,
       page,
       page_number + 1, # Account for zero index
@@ -198,143 +210,6 @@ async def generate_badge_set_images(user:discord.User, all_badges, title, collec
     for page_number, page in enumerate(all_pages)
   ]
   return badge_set_images
-
-# .___
-# |   | _____ _____     ____   ____
-# |   |/     \\__  \   / ___\_/ __ \
-# |   |  Y Y  \/ __ \_/ /_/  >  ___/
-# |___|__|_|  (____  /\___  / \___  >
-#           \/     \//_____/      \/
-def to_thread(func):
-  @functools.wraps(func)
-  async def wrapper(*args, **kwargs):
-    loop = asyncio.get_event_loop()
-    wrapped = functools.partial(func, *args, **kwargs)
-    return await loop.run_in_executor(None, wrapped)
-  return wrapper
-
-@to_thread
-def generate_badge_set_showcase_for_user(user_display_name, page, page_number, total_pages, total_user_badges, title, collected, filename_prefix):
-
-  text_wrapper = textwrap.TextWrapper(width=22)
-  title_font = ImageFont.truetype("fonts/lcars3.ttf", 110)
-  if len(user_display_name) > 16:
-    title_font = ImageFont.truetype("fonts/lcars3.ttf", 90)
-  if len(user_display_name) > 21:
-    title_font = ImageFont.truetype("fonts/lcars3.ttf", 82)
-  collected_font = ImageFont.truetype("fonts/lcars3.ttf", 70)
-  total_font = ImageFont.truetype("fonts/lcars3.ttf", 54)
-  page_number_font = ImageFont.truetype("fonts/lcars3.ttf", 80)
-  badge_font = ImageFont.truetype("fonts/context_bold.ttf", 28)
-
-  badge_size = 200
-  badge_padding = 40
-  badge_margin = 10
-  badge_slot_size = badge_size + (badge_padding * 2) # size of badge slot size (must be square!)
-  badges_per_row = 6
-
-  base_width = 1890
-  base_header_height = 530
-  base_row_height = 290
-  base_footer_height = 200
-
-
-  # If we're generating more than one page we want the rows to simply expand to what's necessary
-  # Otherwise if there's multiple pages we want to have all of them be consistent
-  if page_number == 1 and total_pages == 1:
-    number_of_rows = math.ceil((len(page) / badges_per_row)) - 1
-  else:
-    number_of_rows = 4
-
-  base_height = base_header_height + (base_row_height * number_of_rows) + base_footer_height
-  # base_height = math.ceil((len(badge_set) / badges_per_row)) * (badge_slot_size + badge_margin)
-
-  # create base image to paste all badges on to
-  badge_base_image = Image.new("RGBA", (base_width, base_height), (0, 0, 0))
-  base_header_image = Image.open("./images/templates/badge_sets/badge_set_header.png")
-  base_row_image = Image.open("./images/templates/badge_sets/badge_set_row.png")
-  base_footer_image = Image.open("./images/templates/badge_sets/badge_set_footer.png")
-
-  # Start image with header
-  badge_base_image.paste(base_header_image, (0, 0))
-
-  # Stamp rows (if needed, header includes first row)
-  base_current_y = base_header_height
-  for i in range(number_of_rows):
-    badge_base_image.paste(base_row_image, (0, base_current_y))
-    base_current_y += base_row_height
-
-  # Stamp footer
-  badge_base_image.paste(base_footer_image, (0, base_current_y))
-
-  draw = ImageDraw.Draw(badge_base_image)
-
-  draw.text( (100, 65), title, fill="#8DB9B5", font=title_font, align="left")
-  draw.text( (590, base_height - 113), collected, fill="#47AAB1", font=collected_font, align="left")
-  draw.text( (32, base_height - 90), f"{total_user_badges}", fill="#47AAB1", font=total_font, align="left")
-  draw.text( (base_width - 370, base_height - 115), f"PAGE {'{:02d}'.format(page_number)} OF {'{:02d}'.format(total_pages)}", fill="#47AAB1", font=page_number_font, align="right")
-
-  start_x = 100
-  current_x = start_x
-  current_y = 245
-  counter = 0
-
-  for badge_record in page:
-    badge_border_color = "#47AAB1"
-    badge_text_color = "white"
-    if not badge_record['in_user_collection']:
-      badge_border_color = "#575757"
-      badge_text_color = "#888888"
-
-    # slot
-    s = Image.new("RGBA", (badge_slot_size, badge_slot_size), (0, 0, 0, 0))
-    badge_draw = ImageDraw.Draw(s)
-    badge_draw.rounded_rectangle( (0, 0, badge_slot_size, badge_slot_size), fill="#000000", outline=badge_border_color, width=4, radius=32 )
-
-    # badge
-    b = Image.open(f"./images/badges/{badge_record['badge_filename']}").convert("RGBA")
-    if not badge_record['in_user_collection']:
-      # Create a mask layer to apply a 1/4th opacity to
-      b2 = b.copy()
-      b2.putalpha(64)
-      b.paste(b2, b)
-    b = b.resize((190, 190))
-
-    w, h = b.size # badge size
-    offset_x = min(0, (badge_size+badge_padding)-w) # center badge x
-    offset_y = 5
-    badge_name = text_wrapper.wrap(badge_record['badge_name'])
-    wrapped_badge_name = ""
-    for i in badge_name[:-1]:
-      wrapped_badge_name = wrapped_badge_name + i + "\n"
-    wrapped_badge_name += badge_name[-1]
-    # add badge to slot
-    s.paste(b, (badge_padding+offset_x+4, offset_y), b)
-    badge_draw.text( (int(badge_slot_size/2), 222), f"{wrapped_badge_name}", fill=badge_text_color, font=badge_font, anchor="mm", align="center")
-
-    # add slot to base image
-    badge_base_image.paste(s, (current_x, current_y), s)
-
-    current_x += badge_slot_size + badge_margin
-    counter += 1
-
-    if counter % badges_per_row == 0:
-      # typewriter sound effects:
-      current_x = start_x # ding!
-      current_y += badge_slot_size + badge_margin # ka-chunk
-      counter = 0 #...
-
-  badge_set_filepath = f"./images/profiles/{filename_prefix}{page_number}.png"
-  badge_base_image.save(badge_set_filepath)
-
-  while True:
-    time.sleep(0.05)
-    if os.path.isfile(badge_set_filepath):
-      break
-
-  discord_image = discord.File(badge_set_filepath, filename=f"{filename_prefix}{page_number}.png")
-  return discord_image
-
 
 
 # ________                      .__
