+++ images/templates/badge_sets/badge_set_footer_teal.png
