+++ images/templates/badge_sets/badge_set_row_teal.png
